// backend/src/enums/response.status.code.ts
export enum EnumStatusCode {
  RECOVERED_SUCCESSFULLY = "RECOVERED_SUCCESSFULLY",
  UPDATED_SUCCESSFULLY = "UPDATED_SUCCESSFULLY",
  DELETED_SUCCESSFULLY = "DELETED_SUCCESSFULLY",
  CREATED_SUCCESSFULLY = "CREATED_SUCCESSFULLY",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  NOT_FOUND = "NOT_FOUND",
  BAD_REQUEST = "BAD_REQUEST",
}